:+1: Fixed all warnings

+++ Source/Map.cpp
@@ -17,7 +17,7 @@
 				sf::Vertex tile[2];
 
 				tile[0].position = sf::Vector2f(x, 0);
-				tile[1].position = sf::Vector2f(x, SCREEN_HEIGHT);
+				tile[1].position = sf::Vector2f(x, (float)SCREEN_HEIGHT);
 
 				tile[0].color = grid_color;
 				tile[1].color = grid_color;
@@ -31,7 +31,7 @@
 			sf::Vertex tile[2];
 
 			tile[0].position = sf::Vector2f(0, y);
-			tile[1].position = sf::Vector2f(SCREEN_WIDTH, y);
+			tile[1].position = sf::Vector2f((float)SCREEN_WIDTH, y);
 
 			tile[0].color = grid_color;
 			tile[1].color = grid_color;
